add coname authors license & add test for util functions
